# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + PocketBase backend for QuantumAtlas: paper\ncollection + search + database. Write endpoints and\nthe search/paper read surface require a bearer token\n(PAT or PocketBase session) plus the matching scope.\nSee the auth model docs for the scope vocabulary.",
        "title": "QuantumAtlas API",
        "contact": {
            "name": "QuantumAtlas",
            "url": "https://quantum-atlas.ai"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/devdoc/ticket": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns {url} — a short-lived signed link into the admin-only\ndev-docs site (/devdoc). Opening it sets an HttpOnly cookie and\nredirects to the ticket-free URL; the cookie authorizes the\ndocs for 12h. Admin-only: session token + is_admin or\nis_superadmin on the users record.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Dev-docs ticket",
                "responses": {
                    "200": {
                        "description": "{url}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "dev docs not bundled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Every users record: {users:[{id, name, email, github_login,\ngitea_login, is_admin, is_superadmin, disabled, created,\nupdated}], total}.\nRequires a session AND one of: is_admin, is_superadmin\n(userAdminGuard — PATs rejected; the YAML admin lists are\nonly a boot-time seed for those flags).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "{users, total}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "user management requires admin / PAT auth not accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Body is a JSON object with any of {\"disabled\": bool,\n\"is_admin\": bool} — at least one key required. disabled needs\nadmin-or-above; is_admin needs the is_superadmin flag.\nSelf-protection: nobody may disable\nthemselves or change their own is_admin; admins cannot\ndisable superadmins. is_superadmin is not patchable here.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update user flags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "users record id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated user record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "empty or invalid body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "role too low / self-protection",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003ctoken\u003e\" — two credential shapes\naccepted: a Personal Access Token\n(`Authorization: Bearer qat_...`, minted at\n`/pat` after GitHub OAuth login), or the\nconfig-loaded system PAT (set\n`system_pat.token` in config.yaml on the\nserver, send the plaintext as\n`Authorization: Bearer \u003cvalue\u003e`).\nBrowser callers are authenticated through\npb.authStore (no copy step) — only non-browser\ncallers need an explicit bearer.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}